+++ PSUtil/PSUtil.psm1
@@ -1,10 +1,25 @@
-﻿$script:PSModuleRoot = $PSScriptRoot
-$script:PSModuleVersion = "1.1.5.17"
+﻿$script:ModuleRoot = $PSScriptRoot
+$script:ModuleVersion = (Import-PowerShellDataFile -Path "$($script:ModuleRoot)\PSUtil.psd1").ModuleVersion
 
+# Detect whether at some level dotsourcing was enforced
 $script:doDotSource = Get-PSFConfigValue -FullName PSUtil.Import.DoDotSource -Fallback $false
+if ($PSUtil_dotsourcemodule) { $script:doDotSource = $true }
 
-#region Helper function
-function Import-PSUFile
+<#
+Note on Resolve-Path:
+All paths are sent through Resolve-Path/Resolve-PSFPath in order to convert them to the correct path separator.
+This allows ignoring path separators throughout the import sequence, which could otherwise cause trouble depending on OS.
+Resolve-Path can only be used for paths that already exist, Resolve-PSFPath can accept that the last leaf my not exist.
+This is important when testing for paths.
+#>
+
+# Detect whether at some level loading individual module files, rather than the compiled module was enforced
+$importIndividualFiles = Get-PSFConfigValue -FullName PSUtil.Import.IndividualFiles -Fallback $false
+if ($PSUtil_importIndividualFiles) { $importIndividualFiles = $true }
+if (Test-Path (Resolve-PSFPath -Path "$($script:ModuleRoot)\..\.git" -SingleItem -NewChild)) { $importIndividualFiles = $true }
+if ("<was not compiled>" -eq '<was not compiled>') { $importIndividualFiles = $true }
+	
+function Import-ModuleFile
 {
 	<#
 		.SYNOPSIS
@@ -13,12 +28,14 @@ function Import-PSUFile
 		.DESCRIPTION
 			This helper function is used during module initialization.
 			It should always be dotsourced itself, in order to proper function.
+			
+			This provides a central location to react to files being imported, if later desired
 		
 		.PARAMETER Path
 			The path to the file to load
 		
 		.EXAMPLE
-			PS C:\> . Import-PSUFile -File $function.FullName
+			PS C:\> . Import-ModuleFile -File $function.FullName
 	
 			Imports the file stored in $function according to import policy
 	#>
@@ -28,21 +45,36 @@ function Import-PSUFile
 		$Path
 	)
 	
-	if ($script:doDotSource) { . $Path }
-	else { $ExecutionContext.InvokeCommand.InvokeScript($false, ([scriptblock]::Create([io.file]::ReadAllText($Path))), $null, $null) }
+	if ($doDotSource) { . (Resolve-Path $Path) }
+	else { $ExecutionContext.InvokeCommand.InvokeScript($false, ([scriptblock]::Create([io.file]::ReadAllText((Resolve-Path $Path)))), $null, $null) }
 }
-#endregion Helper function
 
-
-# Perform Actions before loading the rest of the content
-. Import-PSUFile -Path "$PSModuleRoot\internal\scripts\preload.ps1"
-
-#region Load functions
-foreach ($function in (Get-ChildItem "$PSModuleRoot\functions" -Recurse -File -Filter "*.ps1"))
+#region Load individual files
+if ($importIndividualFiles)
 {
-	. Import-PSUFile -Path $function.FullName
+	# Execute Preimport actions
+	. Import-ModuleFile -Path "$ModuleRoot\internal\scripts\preimport.ps1"
+	
+	# Import all internal functions
+	foreach ($function in (Get-ChildItem "$ModuleRoot\internal\functions" -Filter "*.ps1" -Recurse -ErrorAction Ignore))
+	{
+		. Import-ModuleFile -Path $function.FullName
+	}
+	
+	# Import all public functions
+	foreach ($function in (Get-ChildItem "$ModuleRoot\functions" -Filter "*.ps1" -Recurse -ErrorAction Ignore))
+	{
+		. Import-ModuleFile -Path $function.FullName
+	}
+	
+	# Execute Postimport actions
+	. Import-ModuleFile -Path "$ModuleRoot\internal\scripts\postimport.ps1"
+	
+	# End it here, do not load compiled code below
+	return
 }
-#endregion Load functions
+#endregion Load individual files
 
-# Perform Actions after loading the module contents
-. Import-PSUFile -Path "$PSModuleRoot\internal\scripts\postload.ps1"
+#region Load compiled code
+"<compile code into here>"
+#endregion Load compiled code