+++ PSUtil/tests/general/FileIntegrity.Tests.ps1
@@ -28,7 +28,7 @@ function Get-FileEncoding
 	elseif ($byte[0] -eq 0xfe -and $byte[1] -eq 0xff) { 'Unicode' }
 	elseif ($byte[0] -eq 0 -and $byte[1] -eq 0 -and $byte[2] -eq 0xfe -and $byte[3] -eq 0xff) { 'UTF32' }
 	elseif ($byte[0] -eq 0x2b -and $byte[1] -eq 0x2f -and $byte[2] -eq 0x76) { 'UTF7' }
-	else { 'Unknown, possible ASCII' }
+	else { 'Unknown' }
 }
 
 Describe "Verifying integrity of module files" {
@@ -40,11 +40,11 @@ Describe "Verifying integrity of module files" {
 			$name = $file.FullName.Replace("$moduleRoot\", '')
 			
 			It "[$name] Should have UTF8 encoding" {
-				Get-FileEncoding -Path $file.FullName | Should Be 'UTF8'
+				Get-FileEncoding -Path $file.FullName | Should -Be 'UTF8'
 			}
 			
 			It "[$name] Should have no trailing space" {
-				($file | Select-String "\s$" | Where-Object { $_.Line.Trim().Length -gt 0} | Measure-Object).Count | Should Be 0
+				($file | Select-String "\s$" | Where-Object { $_.Line.Trim().Length -gt 0}).LineNumber | Should -BeNullOrEmpty
 			}
 			
 			$tokens = $null
@@ -60,13 +60,13 @@ Describe "Verifying integrity of module files" {
 				if ($global:MayContainCommand["$command"] -notcontains $file.Name)
 				{
 					It "[$name] Should not use $command" {
-						$tokens | Where-Object Text -EQ $command | Should Be $null
+						$tokens | Where-Object Text -EQ $command | Should -BeNullOrEmpty
 					}
 				}
 			}
 			
 			It "[$name] Should not contain aliases" {
-				$tokens | Where-Object TokenFlags -eq CommandName | Where-Object { Test-Path "alias:\$($_.Text)" } | Measure-Object | Select-Object -ExpandProperty Count | Should Be 0
+				$tokens | Where-Object TokenFlags -eq CommandName | Where-Object { Test-Path "alias:\$($_.Text)" } | Measure-Object | Select-Object -ExpandProperty Count | Should -Be 0
 			}
 		}
 	}
@@ -79,11 +79,11 @@ Describe "Verifying integrity of module files" {
 			$name = $file.FullName.Replace("$moduleRoot\", '')
 			
 			It "[$name] Should have UTF8 encoding" {
-				Get-FileEncoding -Path $file.FullName | Should Be 'UTF8'
+				Get-FileEncoding -Path $file.FullName | Should -Be 'UTF8'
 			}
 			
 			It "[$name] Should have no trailing space" {
-				($file | Select-String "\s$" | Where-Object { $_.Line.Trim().Length -gt 0 } | Measure-Object).Count | Should Be 0
+				($file | Select-String "\s$" | Where-Object { $_.Line.Trim().Length -gt 0 } | Measure-Object).Count | Should -Be 0
 			}
 		}
 	}