+++ PSUtil/tests/general/PSScriptAnalyzer.Tests.ps1
@@ -9,18 +9,16 @@ Param (
 
 if ($SkipTest) { return }
 
-if ($env:BUILD_BUILDURI -like "vstfs*") { Install-Module PSScriptAnalyzer -Force -SkipPublisherCheck }
-
 $list = New-Object System.Collections.ArrayList
 
 Describe 'Invoking PSScriptAnalyzer against commandbase' {
 	$commandFiles = Get-ChildItem -Path $CommandPath -Recurse -Filter "*.ps1"
-	$scriptAnalyzerRules = Get-ScriptAnalyzerRule | Where-Object RuleName -NE PSAvoidGlobalVars
+	$scriptAnalyzerRules = Get-ScriptAnalyzerRule
 	
 	foreach ($file in $commandFiles)
 	{
 		Context "Analyzing $($file.BaseName)" {
-			$analysis = Invoke-ScriptAnalyzer -Path $file.FullName -ExcludeRule PSAvoidTrailingWhitespace
+			$analysis = Invoke-ScriptAnalyzer -Path $file.FullName -ExcludeRule PSAvoidTrailingWhitespace, PSShouldProcess
 			
 			forEach ($rule in $scriptAnalyzerRules)
 			{